+++ packages/page-header/CHANGELOG.md
@@ -15,7 +15,6 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ## 0.1.0 (2023-09-06)
 
-
 ### Features
 
-* **mui-page-header:** add PageHeader package ([8401f4b](https://github.com/Availity/element/commit/8401f4ba1ca200824b0542d411ed9d0496bad6ae))
+- **mui-page-header:** add PageHeader package ([8401f4b](https://github.com/Availity/element/commit/8401f4ba1ca200824b0542d411ed9d0496bad6ae))